Change some var types to eliminate more C4244 warnings

+++ MWSE/ArrayUtil.cpp
@@ -38,8 +38,8 @@ namespace mwse {
 		return value;
 	}
 
-	long Arrays::setValue(std::string const& caller, size_t const id, size_t const index, ArrayItem_t const value) {
-		long success = 0;
+	short Arrays::setValue(std::string const& caller, size_t const id, size_t const index, ArrayItem_t const value) {
+		short success = 0;
 		if (id > 0 && id <= arrays.size()) {
 			if (index >= 0) {
 				ContainedArray_t& a = arrays[id - 1];
